+++ packages/patrol/lib/patrol.dart
@@ -2,5 +2,8 @@
 /// flutter_driver.
 library patrol;
 
+export 'src/binding.dart';
+export 'src/common.dart';
 export 'src/custom_finders/custom_finders.dart';
+export 'src/host/host_automator.dart';
 export 'src/native/native.dart';