+++ packages/patrol/lib/src/extensions.dart
@@ -1,13 +1,3 @@
-import 'package:http/http.dart' as http;
-
-/// Provides a method to easily check the meaning of the HTTP status code.
-extension IsResponseSuccessul on http.Response {
-  /// Returns true if the status code is 2xx, false otherwise.
-  bool get successful {
-    return (statusCode ~/ 100) == 2;
-  }
-}
-
 /// Makes it possible to retrieve a name that this [Symbol] was created with.
 extension SymbolName on Symbol {
   /// Returns the name that this [Symbol] was created with.