+++ packages/patrol/lib/src/host/host_automator.dart
@@ -2,30 +2,36 @@ import 'dart:io' as io;
 
 import 'package:patrol/patrol.dart';
 
-typedef _LoggerCallback = void Function(String);
-
 // ignore: avoid_print
 void _defaultPrintLogger(String message) => print('Patrol (host): $message');
+
+/// Configuration for [HostAutomator].
+class HostAutomatorConfig {
+  /// Creates a new [HostAutomatorConfig].
+  const HostAutomatorConfig({this.logger = _defaultPrintLogger});
+
+  /// Called when a host action is performed.
+  final void Function(String) logger;
+}
 
 /// Provides functionality to run actions as the host system (your computer).
 class HostAutomator {
   /// Creates a new [HostAutomator].
   const HostAutomator({
     required PatrolBinding binding,
-    void Function(String) logger = _defaultPrintLogger,
+    required HostAutomatorConfig config,
   })  : _binding = binding,
-        _logger = logger;
+        _config = config;
 
   final PatrolBinding _binding;
-
-  final _LoggerCallback _logger;
+  final HostAutomatorConfig _config;
 
   /// Shortcut for [PatrolBinding.takeFlutterScreenshot].
   Future<void> takeScreenshot({
     String name = 'screenshot_1',
     String path = 'screenshots',
   }) async {
-    _logger('takeScreenshot(name: $name, path: $path)');
+    _config.logger('takeScreenshot(name: $name, path: $path)');
     await _binding.takeFlutterScreenshot(name: name, path: path);
   }
 
@@ -35,7 +41,7 @@ class HostAutomator {
     List<String> arguments = const [],
     bool runInShell = false,
   }) async {
-    _logger(
+    _config.logger(
       'runProcess(executable: $executable, arguments: $arguments, runInShell: $runInShell)',
     );
     final result = await _binding.runProcess(
@@ -45,11 +51,11 @@ class HostAutomator {
     );
 
     if (result.exitCode != 0) {
-      _logger('WARNING: process "$executable" failed');
-      _logger('WARNING:   exit code: ${result.exitCode}');
-      _logger('WARNING:   pid:       ${result.pid}');
-      _logger('WARNING:   stdout:    ${result.stdout}');
-      _logger('WARNING:   stderr:    ${result.stderr}');
+      _config.logger('WARNING: process "$executable" failed');
+      _config.logger('WARNING:   exit code: ${result.exitCode}');
+      _config.logger('WARNING:   pid:       ${result.pid}');
+      _config.logger('WARNING:   stdout:    ${result.stdout}');
+      _config.logger('WARNING:   stderr:    ${result.stderr}');
     }
     return result;
   }