+++ packages/patrol/lib/src/native/native.dart
@@ -1,3 +1,2 @@
-export 'binding.dart';
 export 'contracts/contracts.pbgrpc.dart' show Notification, Selector;
 export 'native_automator.dart';