+++ packages/patrol/test/patrol_tester_test.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:patrol/src/custom_finders/common.dart';
+import 'package:patrol/src/common.dart';
 import 'package:patrol/src/custom_finders/exceptions.dart';
 
 void main() {